Refactor: break out non-React hook helper to calculate Kcl expression value
